refactor(maybeCopyAndPaste): move maybeCopyAndPaste function for better organization
