+++ apps/app/src/lib/query/transformer.ts
@@ -5,7 +5,7 @@ import { settings } from '$lib/stores/settings.svelte';
 import { Err, Ok, type Result, isErr } from '@epicenterhq/result';
 import type { WhisperingError, WhisperingResult } from '@repo/shared';
 import { defineMutation, queryClient } from '.';
-import { maybeCopyAndPaste } from './singletons/maybeCopyAndPaste';
+import { maybeCopyAndPaste } from '../maybeCopyAndPaste';
 import { transformationRunKeys } from './transformationRuns';
 import { transformationsKeys } from './transformations';
